+++ KPM-Backend/Application/Features/Industry/IndustryService.cs
@@ -1,9 +1,8 @@
 namespace Application.Features.Industry;
-using Domain.Entities;
-using Application.Industries.DTOs;
 using Application.Common.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using IndustryEntity = Domain.Entities.Industry;
 
 public class IndustryService
 {
@@ -15,37 +14,41 @@ public class IndustryService
         _logger = logger;
     }
 
-    public async Task<IndustryDTO?> getIndustry(int id)
+    public async Task<List<IndustryEntity>> getAllIndustries()
     {
-        var returned = await _context.Industry
-        .Where(industry => industry.Id == id)
-        .FirstOrDefaultAsync();
-
-        return new IndustryDTO
-        {
-            Name = returned.Name,
-            CreationDate = returned.CreatedDate
-        };
+        return await _context.Industry.ToListAsync();
     }
 
-    public async Task<int?> createIndustry(string name)
+    public async Task<IndustryEntity?> getIndustry(int id)
     {
-        var industry = new Industry
+        return await _context.Industry
+        .Where(industry => industry.Id == id)
+        .FirstOrDefaultAsync();
+    }
+
+    public async Task<int> createIndustry(string name)
+    {
+        var industry = new IndustryEntity
         {
             Name = name,
-            CreatedDate = DateTime.UtcNow,
-            ModifiedDate = DateTime.UtcNow
+            CreatedDate = DateTime.Now,
+            ModifiedDate = DateTime.Now
         };
 
         await _context.Industry.AddAsync(industry);
+        await _context.SaveChangesAsync();
         return industry.Id;
     }
 
     public async Task<int?> updateIndustry(int id, string name)
     {
         var industry = await _context.Industry.FirstOrDefaultAsync(industry => industry.Id == id);
+
+        if (industry is null)
+            return null;
+
         industry.Name = name;
-        industry.ModifiedDate = DateTime.UtcNow;
+        industry.ModifiedDate = DateTime.Now;
         await _context.SaveChangesAsync();
         return industry.Id;
     }
@@ -53,6 +56,10 @@ public class IndustryService
     public async Task<int?> deleteIndustry(int id)
     {
         var industry = await _context.Industry.FirstOrDefaultAsync(industry => industry.Id == id);
+
+        if (industry is null)
+            return null;
+
         _context.Industry.Remove(industry);
         await _context.SaveChangesAsync();
         return industry.Id;
